Moved grpc packages under library/grpc, small optimizations.
