+++ library-enrichment/src/main/kotlin/library/enrichment/gateways/library/grpc/GrpcConfig.kt
@@ -1,4 +1,4 @@
-package library.enrichment.gateways.grpc
+package library.enrichment.gateways.library.grpc
 
 import io.grpc.ManagedChannelBuilder
 import org.springframework.boot.context.properties.EnableConfigurationProperties
