+++ library-enrichment/src/main/proto/book_response.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 option java_multiple_files = true;
-option java_package = "library.enrichment.gateways.grpc";
+option java_package = "library.enrichment.gateways.library.grpc";
 option java_outer_classname = "BookResponseProto";
 
 message BookResponse {
