+++ library-enrichment/src/main/proto/book_update.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 import "book_response.proto";
 
 option java_multiple_files = true;
-option java_package = "library.enrichment.gateways.grpc";
+option java_package = "library.enrichment.gateways.library.grpc";
 option java_outer_classname = "BookUpdateProto";
 
 // The definition of Book update service.
